+++ src/main/java/mybatis/mapper/users-mapper.xml
@@ -3,6 +3,9 @@
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="users">
+	<select id="findById" resultType="data.user.User" parameterType="java.lang.String">
+		SELECT * FROM USERS WHERE ID= #{ID}
+	</select>
 	<insert id="create" parameterType="java.util.HashMap">
 		INSERT INTO USERS(ID, PASS, NICK) VALUES(#{id}, #{pass}, #{nick})
 	</insert>
