+++ src/main/java/user/controller/UserJoinController.java
@@ -1,4 +1,4 @@
-package controller.user;
+package user.controller;
 
 import java.io.IOException;
 
@@ -8,10 +8,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/index")
-public class IndexController extends HttpServlet {
+@WebServlet("/user/join")
+public class UserJoinController extends HttpServlet {
+	
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/WEB-INF/views/index.jsp").forward(req, resp);
+		req.getRequestDispatcher("/WEB-INF/views/user/join.jsp").forward(req, resp);
 	}
 }
