+++ src/main/java/user/controller/UserLoginController.java
@@ -1,4 +1,4 @@
-package controller.user;
+package user.controller;
 
 import java.io.IOException;
 
@@ -8,11 +8,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/user/join")
-public class UserJoinController extends HttpServlet {
-	
+@WebServlet("/login")
+public class UserLoginController extends HttpServlet  {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/WEB-INF/views/user/join.jsp").forward(req, resp);
+		req.getRequestDispatcher("/WEB-INF/views/user/login.jsp").forward(req, resp);
 	}
 }
